requirements.txt
streamlit
python-whois
pytz
pandas
requests